Add run-relayer command (#11)
* Add run-relayer command

Adds a oz-gsn run-relayer command. It downloads the relayer binary from a github release to the home folder, deploys the relay hub contract, starts the relayer process, registers the relayer on the hub, and funds it. It can be called with a `--detach` flag so it exits after the setup is ready, so it can be used for test setup.

* Rename process to subprocess

* Do not accept both port and relayUrl opts

* Use env-paths cache path instead of homedir for relayer bin